If Statements on Checks
Added really unsightly if statements to lock the user into specific
input types. The only if statement that does not work as intended is the
Zip Code check as it does not allow for hyphens, but allows any amount
of spaces. I will leave as is until I learn a better system to account
for all Zip Codes across the world.